+++ database/seeders/DatabaseSeeder.php
@@ -4,6 +4,7 @@ namespace Database\Seeders;
 
 use App\Repositories\DatabasesRepository;
 use App\Repositories\FrameworksRepository;
+use App\Repositories\ProgrammingLanguagesRepository;
 use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
@@ -17,5 +18,6 @@ class DatabaseSeeder extends Seeder
     {
         DatabasesRepository::factory()->create();
         FrameworksRepository::factory()->create();
+        ProgrammingLanguagesRepository::factory()->create();
     }
 }